alx_book_store added

+++ alx_book_store.sql
@@ -34,7 +34,7 @@ CREATE TABLE Order_Details(
 	order_detail_id INT PRIMARY KEY,
 	oder_id INT,
 	book_id INT ,
-	quantity FLOAT,
+	quantity DOUBLE,
 	FOREIGN KEY (order_id) REFERENCES Orders(order_id),
 	FOREIGN KEY (book_id) REFERENCES Books(book_id)
 )
